app/requirements.txt
Flask
uwsgi
requests
ShopifyAPI
